Move installation monitor to installation list with toggle button
