+++ client/galaxy/scripts/components/Toolshed/Index.vue
@@ -1,8 +1,5 @@
 <template>
     <div class="overflow-auto h-100 p-1" @scroll="onScroll">
-        <p>
-            <Monitor @onQuery="setQuery" />
-        </p>
         <div v-if="error" class="alert alert-danger">{{ error }}</div>
         <div v-else>
             <b-input-group class="mb-3">
@@ -24,19 +21,17 @@
                 <SearchList :query="query" :scrolled="scrolled" @onQuery="setQuery" @onError="setError" />
             </div>
             <div v-else>
-                <InstalledList :filter="queryInput" />
+                <InstalledList :filter="queryInput" @onQuery="setQuery" />
             </div>
         </div>
     </div>
 </template>
 <script>
-import Monitor from "./Monitor";
 import SearchList from "./SearchList/Index.vue";
 import InstalledList from "./InstalledList/Index.vue";
 
 export default {
     components: {
-        Monitor,
         SearchList,
         InstalledList
     },
