+++ client/galaxy/scripts/components/Toolshed/InstalledList/Monitor.test.js
